[SDR-357] refactor: Menu 컴포넌트 KebabMenu로 이름, 경로 수정
